Split ProbingTermination into PaddingTermination and DuplicateTermination.
